+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 ## Added
     - Execution de npm init une fois la configuration validé
     - Ececution de l'installation de dependances
+    - Execution de l'installation de pnpm
 ## Fix (Thanks for your feedback Mahery Randrianirina ;)
     - Ajout de l'initialisation automatique du module Go (go mod init) si le fichier go.mod est absent.
     - Installation automatique des dépendances via go mod tidy lors du make build.